pyproject.toml
[build-system]
requires = ["flit_core >=3.2,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "RETHINKING_GENERALIZATION"
version = "0.0.1"
description = "implementing famous research paper: Understanding Deep Learning Requires Rethinking Generalization"
authors = [
  { name = "KaanKazguc" },
]
license = { file = "LICENSE" }
readme = "README.md"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License"
]
requires-python = "~=3.13.0"
dependencies = [
    "torch>=2.1.0",
    "torchvision",
    "pandas",
    "Pillow",
    "numpy",
    "matplotlib",
    "opencv-python",
    "python-dotenv",
]


[tool.ruff]
line-length = 99
src = ["RETHINKING_GENERALIZATION"]
include = ["pyproject.toml", "RETHINKING_GENERALIZATION/**/*.py"]

[tool.ruff.lint]
extend-select = ["I"]  # Add import sorting

[tool.ruff.lint.isort]
known-first-party = ["RETHINKING_GENERALIZATION"]
force-sort-within-sections = true

[tool.pyright]
extraPaths = ["RETHINKING_GENERALIZATION"]
